Checkpoint for a more extensive refactoring
This is going to fail, but it can't be fixed just through configuration. First, there's the polyLifeCycle error. Then, there's the error of mixing mocha and jest, which is simply not possible. Despite possible conflicts afterwards, I'm going to go back and refactor this to using jest before merging, so that at least I got one of the problems solved