PROD4POD-807 - Run podSpec against pod.js
Since we are running the podSpec tests against RemotePod, and even DefaultPod
(which is unused aside from some tests), it also makes sense to run them against
the BrowserPod implementation in pod.js.

I opted for not including dummy-server for now, even though the DefaultPod tests
(still) do. That was needed back when we still had the polyOut.fetch API, and
tests for it. At the moment we don't need it, though we may need it again later.