Clear IndexedDB before test
This is a pretty crude fix for the test failures we're seeing. From my
research, it appears that while Cypress does clear localStorage before
each test, it does not do the same for IndexedDB, so we have to do it
manually.

The failing test, in particular, only works with a clean database, since
it's testing first run behaviour.

There's definitely a more elegant way to do this, e.g. a custom Cypress
command instead of a simple async function, but this does the trick for
getting things green again, for now.