[Task] {PROD4POD-1445} Unify API in a single package... Starting with communication (#763)
* Initial moving

* Got it to build :fire:

* :see_no_evil:

* :rotating_light:

* Moved bubblewrap completely

* :memo: :sparkles:

* Change line :rotating_light:

* That package no longer exists

* Adds index

* Single point of export

* Recomposing the build

* Recomposing deps

* Fixes and syncs lockfiles

+ unneeded meta info

* Temporary rearranging deps

* Accomodate incorporation of new module

* Change path to bubblewrap Java stuff

* :rotating_light:

This wasn't changed from before, but it had somehow evaded linting before. Also, some changes in format were made, so disabling some errors on the next line stopped making sense

* Deleting no longer needed files :coffin:

And updating configuration for new module

* Move :memo:

* Removed directory, hoisted up

* :bug: path

* Adjust config

* Solving most of them

Preparing a PR on common code

* :rotating_light:

* Adapt package to current temporary situation

* :coffin:

* Definitely eliminate Java from this moved repo

* Really, really, eliminate them

* :recycle: completed

* :recycle: paths

* Still solving merge

* Migrating and fixing build

* We weren't running these anyway

* :white_check_mark:

* Adapts :memo:

* :coffin:

* Improve configuration

* Finishes incorporation

* Downgrading node-fetch version for :white_check_mark:

* Putting that into the configuration

* Got it to build

* Got it to build and test

But it issues a series of warning errors about this not defined

* :rotating_light:

* Retire from packages

* Remove traces of moved module

* Need to downgrade @rdfjs/dataset because version 2 breaks build

* Checkpoint for a more extensive refactoring

This is going to fail, but it can't be fixed just through configuration. First, there's the polyLifeCycle error. Then, there's the error of mixing mocha and jest, which is simply not possible. Despite possible conflicts afterwards, I'm going to go back and refactor this to using jest before merging, so that at least I got one of the problems solved

* :coffin:

* Solving should go before the commonjs() declaration

This does not solve anything, it's just good practice

* Managing to use dynamic imports

* Changed to dir, kept only one of the formats

* :rotating_light:

* More already hoisted configuration

* Eliminate from list of packages

* Exclude paths not tested with jest

* Last ditch attempt

* Eliminated due to impossibility of testing it alongside jest

* Additional :coffin:

* Simplify import, addressing nits

* Eliminates :coffin:

Please note that we no longer can create an IIFE, due to the fact that we have multiple, dynamically imported modules

* :memo: addresses nits

* Moved READMEs to their own directories, left only a stub at the top

Just a reminder that this is provisional, and needs to be merged with the api part soon-ish