Automatically build the Android core bindings
While at it, I:

1. Added a non-phony target for the Android libraries, for incremental
builds. However, it's a bit brute force, doing anything more fine
grained or less redundant is tricky when the actual Android build logic
is in the separate build_for_android.sh script. But seems good enough
for now.

2. Renamed the "android_libraries" phony target to "android_bindings",
because ultimately that's what these things are: bindings.

3. Added an anaolous "ios_bindings" phony target for convenience.