Fix pod.js generation in lexicon feature
It was previously skipping the pod.js copy, because the 'public'
directory already existed - it happened to work with our other features
because pod.js was going right into the build directory ('dist') there,
and the genPodJs plugin happened to be the first one to be called.

I believe the logic is now what it was supposed to be originally -
merely ensuring that the directory pod.js is supposed to be copied into
actually exists.