Revert "Fix zip file reading via pod.js"
This reverts commit 8082a9e4b9aca2ba9b3da32c023f6bb8dd90af3a.

Since we expose both "id" and "getId" on Android and iOS, it seems
relatively futile to _try_ and avoid using "id" when it's readily
available. I think we ought to go one way or another: Eliminate "id" or
"getId" (and the same for all the other accessors and properties).

For now, a better fix is for podjs to replicate the behaviour we see on
Android and iOS.