Cache totally eliminated
Since performance of the MacOS runner is dismal, it takes an inordinate amount of time to unpack the cache and to compute the hashes. So I think overall we're better off eliminating it. Installing the modules is I/O bound, and it's not that bad in that area, so we might actually see our workflows run *faster* without the cache