[🐛] {PROD4POD-1779} Try to fix lockfile-hashing timeout (#1077)
* :arrow_up: cache version

And add self path to trigger workflow

* Cache totally eliminated

Since performance of the MacOS runner is dismal, it takes an inordinate amount of time to unpack the cache and to compute the hashes. So I think overall we're better off eliminating it. Installing the modules is I/O bound, and it's not that bad in that area, so we might actually see our workflows run *faster* without the cache

* While I was at it, :memo: fixed

We're no longer testing Android here

* :arrow_up: for the action, as well as node

Well, we can use it to see if there's any hiccup with the latest version