Eliminate copying the test feature to the device
# ✍️ Description

At the moment, every time we deploy the tests, the build process will
copy over the feature named "test" to the device.

Now that we're always bundling - but not displaying - it, we can just
add a preference for forcing it to show up, and execute the tests in
it that way, so we don't need this logic anymore. It's sometimes
causing trouble when trying to run the tests - happened to me again
just now - and it's not worth debugging if we don't need it.

Related to PROD4POD-1721.