(Exploratory) eliminating ownership transfer
This code was introduced more than three years ago here 9a3bcb93d71bf7feb856167e598eae591bcdeedd I haven't found any evidence of someone receiving the ownership transfer and actually using it. It's got no tests or specs, so I'm not sure it serves any purpose, and the use of unrestricted targets is a potential security issue. It's probably this will trigger a test failure, and in that case I'll be a bit more acquainted with what this really does. Even if it does not, it will have to be checked in practice. At any rate, we'll understand better what some essential part of the code really does