[🐛] Suppress only the warnings we're interested in (#1146)
# ✍️ Description

The code has been DRYed. I've created a better regex, because the previous one, as indicated by @snoack in #1126, was eliminating every warning.

This one takes care of the two kinds of warnings that were there

## ℹ️ Other information

There are a couple of places where the same pattern was used to suppress warnings.  If this new code is OK for everyone (it was already used in a couple of places), I can also refactor those within this same PR.

♥️ Thank you!

I'll extend this change to other rollup configurations in other PRs