PolyLook: Add tokens for themes and buttons (#1139)
# ✍️ Description
- [x] **Create "coop" theme in PolyLook** and added in Membership feature with a button as an example. 
        `--theme-primary-color` is the background color, since that variable is already used in the Screen component.
        `--theme-secondary-color` is the text color
        `--theme-tertiary-color` is `--color-red-300` for the three themes that we already have.
- [x] **Add button tokens and implement them in PolyButton**. New variables:
        `--pl-button-color-fill-primary`
        `--pl-button-color-text-primary`
        `--pl-button-color-fill-secondary`
        `--pl-button-color-border-secondary`
        `--pl-button-color-text-secondary`
- [x] ~~**Fix button types in PolyButton**: "primary" and "secondary" buttons, instead of "filled" and "outline".~~
     
### 🏗️ Fixes [PROD4POD-1861](https://jira.polypoly.eu/browse/PROD4POD-1861)
### 🏗️ Fixes [PROD4POD-1869](https://jira.polypoly.eu/browse/PROD4POD-1869)

## Other information 
FIGMA DESIGNS - these are provisional, but we can have an idea of the color scheme: https://www.figma.com/file/H4JnISIsC06WPouox7CITh/MAIN-pod-GDPR-helper-feature?node-id=1286%3A4364

There is something that I was not sure how to address here and I am open to suggestions of improvement. 

As you can see in the Figma mock-up, there are two background colors in the Membership feature: light-red and white. In order to implement the light-red un the onboarding screen, I have created the new `coop-theme`. The rest of the feature would use the `light-theme`, and the primary-buttons type would be `medium` (or "primary-medium") in order to change the "fill" color to the red-300 - same as it is done in the `import` view in the Google feature. 

The issue here would be that the `primary-medium` won't be using the respective button tokens:
```
.poly-button.primary-medium {
  background-color: var(--theme-tertiary-color);
  color: var(--color-indigo-950);
}
```

Another option would be to create another theme, something like `coop-home-theme` for the light-red background and `coop-theme` for the white background and we could set the different button variables there for the white-background/red-button combination.

Please, let me know which solution you think it's the best one. ❤️