PROD4POD-1884 - Fix Screen component overlap (#1154)
# ✍️ Description

Every time we implemented the `Screen` component in a new feature, it covered the nav bar on the top. We have been fixing it in each feature, but it happened every time that we created a new one. 

In order to create a more permanent solution, I have created a `feature-container` class in `poly-look` and added it to the "feature" and "preview-feature" templates in `poly-cli`. This fixes the problem for potential new features.

I have also added/replaced that class for the current features that didn't have it yet.

### 🏗️ Fixes [PROD4POD-1884](https://jira.polypoly.eu/browse/PROD4POD-1884)


## ℹ️ Other information

[How it looked like.](https://jira.polypoly.eu/secure/attachment/14665/How%20it%20looks%202.png)
[How it looks now.](https://jira.polypoly.eu/secure/attachment/14663/How%20it%20should%20look%202.png)