send error reports to new server endpoint - PROD4POD-943 (#1132)
## ✍️ send error reports to new server endpoint - PROD4POD-943

We now can send error reports to our new endpoint `/errorReports`. User needs to consent for this upload to our servers.

To do this, we add an `Endpoint.uploadToServer()` to handle the upload of the error logs, that is called when dialog button is selected. 
The text of button has been adapted precisely [screenshot below]
Implementation both for android and iOs.

Some other small refactorings on the go.

Fixes ([PROD4POD-943](https://jira.polypoly.eu/browse/PROD4POD-943))