[iOS] Encode values automatically (#1115)
# ✍️ Description

## What is the problem? 

Exposing APIs from Core is cumbersome. These are the steps involved:
1. Create a new function in the C interface. Decode arguments passed from Swift manually. 
2. Create a corresponding function in Swift where the values from and to Rust are encoded/decoded manually. 
3. Repeat the steps above for java interface. 

It is very time consuming and it can be done a lot better. 

Here is what can be done to improve: 
- [ ] No more need to expose new API in C and Java interface. This can be achieved by having a single function in the interface that handles requests from core, encodes/decodes values, and passes them to the Rust core. 
- [x] Automate Encoding iOS
- [ ] Automate Decoding iOS
- [ ] Automate Encoding Android
- [ ] Automate Decoding Android

I aim to address these improvements, as a series of self-contained pull requests. This PR is the first in this series and addresses "Automate Encoding iOS". 

## What was done in this PR? 

I automated the conversion from `Swift Type` to `MessagePackValue` which is then converted to bytes. 

Previously one had to go through each value and encode it manually, which if you have a big deep map, can be very cumbersome. 

There is one disappointment that I have. I expected the encoding of `PlatformRespose` to be understood by the Rust decoder. Alas, that is not the case, so I had to provide specific encoding instructions.

Co-authored-by: Ghenadie Vasiliev-Pusca <ghenadie.pusca@polypoly.net>