Fix google importer - Use library to parse dates (#1152)
# ✍️ Description

I got rid of the regexes that we were using for parsing dates. Now we
are using `dayjs` to parse.

There is a small complication though. dayjs does not support named
offsets like `EEST` so before I parse the dates I need to convert the
`EEST` to UTC.

For that I scraped https://www.timeanddate.com/time/zones/ and made a
json mapping named offsets to UTC.

Co-authored-by: Ghenadie Vasiliev-Pusca <ghenadie.pusca@polypoly.net>