[♻] {PROD4POD-1889} Reinstate instrumentation tests (#1159)
# ✍️ Description

Changes the workflow from scheduled to triggered by certain paths. I
haven't installed it back in the iOS workflow for a couple of reasons.
It would require adding actions for selecting some paths, and then
checking for them as we do in the Linux tests; it would also add to the
time required to run that. By having it in its own workflow, we need not
wait (if we don't want to), but we will still get what we were looking
for, errors for specific pushes. By narrowing paths, we will also be
able to not have to wait for basically every push.

The modules cache has been disabled, too. Restoring the cache took as
much as downloading and building the modules. It was already eliminated
in the other iOS workflow, so I guess it's just as well we do it here
(saving a couple of minutes)

## 🏗️ Fixes PROD4POD-1889

By making it run triggered by pushes


♥️ Thank you!