feat(polyLook) - prod4pod-1879 add icon options to PolyButton (#1168)
# ✍️ Description
This PR adds icon(s) to the PollyButton component.

While in practice we would only use one of the icon options, there are
no code restrictions for that.

Initially I had planned for `iconLeft` and `iconRight` to accept either
a string representing a path or a component as prop (current
implementation) and let the button decide how to render that.
After some consideration to theming I've decided to get rid of the `img`
support and keep only the component as prop approach. That's done to
encourage the inline SVG approach to icons.
I haven't added any css rule for the fill because I'm not sure how the
icons would look like, so I'll leave it up to the dev using this
component to decide.

<!-- Link here the related JIRA issue in case of a bug / feature -->
### 🏗️ Fixes
[PROD4POD-1879](https://jira.polypoly.eu/browse/PROD4POD-1879)



## ℹ️ Other information
In the examples below, the square is an inline SVG with fill set to
`currentColor`. For theming that would be the ideal approach.


![icon_all](https://user-images.githubusercontent.com/17316907/188896028-70fbe019-3026-49b7-b49e-e1e729726a10.JPG)