Define Core api with data models (#1166)
# ✍️ Description
Refactored `core` public API, so it's "functional" API is expressed
through data models. Instead of having `n` functions in the API there is
just one function which accepts an Enum with `n` cases. This design
helps with keeping the bridge API(C and Java interfaces) stable, while
the rest of the heavy lifting is done in Rust/Swift/Kotlin.

What changed:
- Define the single API function `execute_request` which is exposed to
Platforms.
- Defined the available `requests` in Core and on the platforms.
Performed the proper mappings.
- Cleaned up `c_interface` and `java_interface`.
- Removed `cbindgen` for generating the C header, as it is not needed
anymore. The C header now should change only when there is a new channel
of communication.
- Did split the `core/lib/src/lib.rs` into multiple files, for better
separation of concerns.
- Polished iOS and Android Core API's.

Co-authored-by: Tim Palade <palade.timotei@gmail.com>