[🐛] {PROD4POD-1805} Changed paths to new name (#1173)
# ✍️ Description

And factored out name to not repeat it. We could possibly do the same
with other paths that are repeated in the importer, but that's out of
scope for the time being.

Changed also in the description of the download, which is possibly used
to generate reports.

Since some customers might want to re-analyze the old files, they might
still have the old `pages` directory. All the files have been
replicated, hoping this will be enough to make it work. Since both work
on the same area of the facebookAccount, one or the other should get it.
I'm not so totally sure the "disabled" mechanism will work correctly, it
might depend on the order of applying it... but it was the best I could
come up with.


### 🏗️ Fixes PROD4POD-1805

By changing the directory name
![Captura de pantalla de 2022-09-13
08-19-11](https://user-images.githubusercontent.com/500/189825049-643b0a66-41e6-449a-91e6-16e95b252cf2.png)

Tested with a fresh download with my data. 

## ℹ️ Other information

Stuff to do (out of scope)

* Change example downloads
* Check other directories


♥️ Thank you!