Tweak GitHub Actions (#1186)
# ✍️ Description

Now, additionally to the tests that are run when a push happens, it runs
additional tests when certain things happen in a PR. This is not really
necessary, since PRs follow the same code path that would follow the
latest commit. Duplicated tests can be useful to detect flappers, but
this specific test does not have a lot of them. Will leave just one
duplicated test to take care of that, and that's that.

♥️ Thank you!