Refactoring Importer Framework (1): Setting the base (#1190)
# We want to change the importer framework to allow it to use rdf and
improve the inner workings a little while on it

### What we want to achieve:

- Neither importers nor analysis alter the dataAccount in their scope ->
They return the result and status report on how it went
- Importers can be extended to implement rdf
- RDF importers can return functions instead of data to support lazy
loading

### What we need to do:

- Change the importers from altering dataAccount in scope to returning
result and report as response. Result before was used to refer to
report, but now describes the payload that is returned in case of
success
- Add a RDFImporter BaseClass and change the Importers that want to have
rdf to using it
- Change the importers one by one to fit the new schema(*)


(*) I did not want to make this a monster PR so I entirely supported the
old architecture and have them running side-by-side until this is
finished. This means TODO comments on having to remove this later. Some
parts (eg. `runOutdatedImporter`) are only kept for testing and are not
used in the actual app. This allows us to not have to change everything
at once, while keeping the app running.

<!-- Link here the related JIRA issue in case of a bug / feature -->
### 🏗️ Fixes PROD4POD-1904 (1/n PR)