[♻] Add tests to bubblewrap (#1142)
# ✍️ Description

After the error detected after merging #1133, it's probably likely that
some tests are needed to avoid changing the interface, since those
changes are not really caught in app-level tests. This is an attempt to
do so. It checks that the classes we know are going to go through
remote-pod are, at least in number, the same. It also checks that the
declared classes can roundtrip (although for the time being there's a
single test, should have more before merging).

Some refactoring on the fly has been made, mainly factoring out some
constants (and exposing them as part of the interface)

## ℹ️ Other information

I'm adding this expecting that it adds some value to the code base and
it helps catch errors in the future. Also, as usual, to become more
acquainted with that said codebase so that it's easier for us to make it
evolve.

♥️ Thank you!

Co-authored-by: irstavr <irini.stavrakantonaki@polypoly.net>