[:recycle:] `changeListener` moved to the constructor (#1210)
# ✍️ Description

Assigning a value to an attribute violated encapsulation, which is easy
to do in JS, but still. This is a cleaner API, with a private
`changeListener` which is assigned when the object is built.

### 🏗️ Works with PROD4POD-1931

Still cleaning up and refactoring the importer, looking at improving
coverage. Related to #1209 and #1204, but wanted it to be considered on
its own.

Also it's a change of API, which is better if considered on its own.