[:recycle:] Eliminate unneeded coupling with zipFile :coffin: (#1219)
# ✍️ Description

Again, this was found wanting for #1218. This coupling was probably not
needed, so that all information for an entry occurs in the entry itself,
and it does not need to be handled by the hosting zipFile. Since this is
a Mock, we don't really need more than that.

## ℹ️ Other information

Might include some tests, and additional spin-off of all ZipMockFile
methods that create Entries.