[:recycle:] {PROD4POD-1940} `polyLifecycle` is probably unused (#1212)
# ✍️ Description

First hint is the fact that it's hidden in the documentation, but then
their functions are not invoked anywhere or used in production,
apparently. Its behavior is not documented, and its code has been
sitting there for some time.


### 🏗️ Fixes
[PROD4POD-1940](https://jira.polypoly.eu/browse/PROD4POD-1940)

> I created this one, but still

## ℹ️ Other information

I'll proceed cautiously here, since it might be used under some name
we're not familiar with. So consider this a draft, although feedback is
welcome. Anyway, this code was created by [this
commit](https://github.com/polypoly-eu/polyPod/commit/2ef98b4b836cd4fae1d9b274bab1ca33598dffcc),
which says `experimental`

WIP:
* [x] Eliminates `listFeatures`
* [x] Eliminates polyLifeCycle as a pod Property. 
* [x] Totally eliminates PolyLifeCycle as a class 

 ♥️ Thank you!