[:coffin:] Refactoring return values (#1209)
# ✍️ Description

While doing #1204 a bit of refactoring is needed; and I realized this
return value is never used. Giving it its own PR since it's a change in
API, and want it to be tested with the rest.

Also, some refactoring mainly eliminating `await` and `return` where
they're not actually needed.

### 🏗️ Fixes PROD4PROD-1931

It contributes to cleaning the repository, and improve test coverage in
`importer`; eliminating code that's not used will no doubt contribute to
that.