Move all documentation in one central point - Prod4pod 1937 (#1231)
## Scope

In this PR, we provide one central point for all the documentation
generated in polyPod.
Now all docs are located under `/docs` folder in a structure similar to
the project's.
e.g. the docs of `api` can be found under
`/docs/platform/feature-api/api/' dir

To do so, we:
Added a new script to `build.js` to generate all the documentation. This
new script reads the subdirectories which already have `doc` script and
executes the `jsdoc` or `typedoc` to generate our docs.
 
As the documentation till now was partly working only for .js and .ts
files, we installed the `better-docs` tool of `jsdoc` in order to
support the `jsx` files for the react components.
Therefore, we needed to set a `jsdoc.conf.json` under every dir that has
both .js and .jsx files to support them both.
And we corrected all the scripts to make use the config file: `"doc":
"jsdoc -c jsdoc.conf.json"`

 ### To remark:
As a reminder for any new code that will need documentation, we'll
require to copy the `jsdoc.conf.json` to that dir as well. The only
change required is to change the 'name' of the documentation page and
the _output_ folder:
e.g. for 'poly-look' we have:
``` .json
    "templates": {
      "better-docs": {
        "name": "_poly-look documentation_" // TBC
      }
    },
    "opts": {
      "template": "../../node_modules/better-docs",
      "destination": "../../docs/feature-utils/poly-look",  // TBC
      "recurse": true,
      "verbose": true,
      "readme": "README.md"
    }
```

Also, for the 'prettiness' of the resulted page, we set the Readme as
the global html page, the param "readme" on the script reads the path of
the file. In case it's missing, the script will fail. (Thus, a copy of a
'Readme' was added to google feature as there was none till now).

<!-- Link here the related JIRA issue in case of a bug / feature -->
### 🏗️ Fixes
([#Prod4Pod-1937](https://jira.polypoly.eu/browse/PROD4POD-1937))