[🐛] [PROD4POD-1959] Updates privacy policy updating script (#1227)
# ✍️ Description

Initially: ~~For some reason, the google console was requesting a
privacy policy for the `ro` locale. Maybe it was detecting downloads
from that area. This is an MVP which fixes the error by creating a file,
originally copied from the English version. Of course, it will need to
be translated.~~

I've added the Danish privacy policy (also in English, I'm afraid) and
adapted the script that downloads them to the new URL. Not run it for
the time being, since there seems to be a bug in the text. The other
privacy policies have been updated to their latest version.

### 🏗️ Fixes 1959 partially