[♻] Add test for bubbles, refactor to avoid content error (#1220)
# ✍️ Description

This code was the only one using the construction `zipFile.getContent`,
which apparently disappeared a some point in time. It's been refactored
to work as intended, getting the content of a single `ZipEntry`. Adds a
tests which fails if the original content was used.

### 🏗️ Fixes PROD4POD-1940

This was found while doing #1219, so it's a blocker for that one.

## ℹ️ Other information

Deep tests have not really been performed; they're mainly sanity tests.
Rendering tests are not checked either. Better mocks would probably be
needed for better tests.

(Unrelated) coupling with `zipFile` has been eliminated from the
function, since it's no longer needed. The function was returning a data
structure which included it, not doing that does not seem to affect the
outcome.