correct jsDocs following changes in #1231 (#1233)
## Scope

Fixes the jsDocs mostly in react components (and in some `.js`),
following the new structure of `/docs/documentation` in
https://github.com/polypoly-eu/polyPod/pull/1231
so that documentation is generated without any errors.


## 🧪 How to Test 

Just need to run `./build.js doc` and check the generated documentation
under `/docs/documentation/` folder.


## 📖 As future reference of jsDoc usage:

To annotate the params of a component: `@param {type} name -
description`
For an _optional_ param should be: `@param {type} [name] - description`

Potential `{type}` for: 
a general object:`Object`
string value:`string`
boolean value:`boolean`
array of objects:  `Array.<{type}>`
etc. 
or if we know the exact type, we can link to its class directly.

When the param is an object with different fields, should be _first_
annotated as:
`@param {Object} props`
and the rest children fields carry its name, so they're linked to it:
`@param {string} props.firstField - the first field of the object`

For a `callback` type we can firstly define it:
```
/**
 * Callback for X.
 * @callback nameCallback
 * @param {type} nameCallback - callback when ...
 */
```
and use it:
`@param {nameCallback} myCallbackParam - onClick function`


⚠️ also in the case of `export const` callbacks should be annotated
specifically with `@callback <itsName>` (so its params are included into
the documentation)
e.g. 
```
/**
 * It renders an empty chart.
 * @callback MyChart
 * @returns {JSX.Element} an empty JSX div
 */
export const MyChart = () => {
 ...
  return <div></div>;
};
```

A general example of a `ClickableCard`:
```
@component
@param {Object} props
@param {JSX.Element} props.children JSX elements displayed inside the card
@param {onClickCallback} props.onClick onClick function
@param {string} props.buttonText displays a button with the string at the bottom if passed
@param {string} [props.onClickEvent] if a button is active, makes the button the only clickable part. Defaults to false. 
@returns {JSX.Element}  A function that returns a div
 */
const ClickableCard = ({
  children,
  onClick,
  buttonText,
  onClickEvent = false,
}) => { (edited) 
```

For more detailed usage of jsdocs, please, consult the official
[jsdoc](https://devdocs.io/jsdoc/) documentation :)