feat(fbImport) - prod4pod-1547 use SideSwiper and related components for HTRT (#853)
This PR finally adds the SideSwiper to facebookImport.

A few notes:
- there are some changes in poly-look that I'm including in this PR.
Normally I would have tried to split them out and have a separate PR,
but for the sake of speed and because it's mostly style related I'm
leaving them here.
- I've shuffled a few things around: `BaseInfoPopUp` is now the main pop
up component and it handles what content to show based on info from
context; Context is only responsible for storing the pop up name. And we
still set the pop up using `InfoButton`. I'm a bit tempted to move
`InfoButton` to `popUps` just for the sake of keeping as much of the
logic co-located, but I would like some feedback first as moving files
around is a bit annoying.

Co-authored-by: mteresa-jimenez <mteresa.jn@gmail.com>