PROD4POD-1972 - Eliminate PolyUri etc from the API
All of this code was used to essentially share two trivial lines of code
between pod.js and MockPod. Considering that there is no requirement for
polyPod URLs to have exactly that protocol, I got rid of it. This got
rid of four exported symbols that are irrelevant for users of the
feature API.