Move our own RDF types into a dedicated namespace
We did use "RDF" quite a lot for the types imported from the "rdf-js"
package, so I changed that to "RDFJS".

The main benefit of this change is that the source code documentation of
the API package is now no longer being bloated with RDF-specific types
at the top level.