:bug: sanitizes warning by removing newlines (#1244)
This might or might not be enough for
https://github.com/polypoly-eu/polyPod/security/code-scanning/66. Let's
see

## Scope

There are issue revealed by the CodeQL security scanning policy. This is
limited only to those with high priority.
This limited approach might or might not solve the problem; a more
extensive refactorization of logging might be needed. We'll get there if
needed.

## 🏗️ Implementation 

"Log injection" starts with using a newline character to kinda "put" new
lines into the log. This could affect mainly logs sent to us by users, I
guess, which might cause some problems if it's done with malicious
intent. At any rate, maintaining some level of health is desirable.